added test for new report
